+++ src/main/java/com/electrocucaracha/apps/cdp/models/HasId.java
@@ -1,5 +1,5 @@
 package com.electrocucaracha.apps.cdp.models;
 
 public interface HasId {
-	public int getId();
+	public long getId();
 }
